[TEST] 인수 테스트 추가 #149
