+++ src/test/java/com/sports/server/query/acceptance/LeagueQueryAcceptanceTest.java
@@ -1,24 +1,23 @@
 package com.sports.server.query.acceptance;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-
 import com.sports.server.query.dto.response.LeagueDetailResponse;
 import com.sports.server.query.dto.response.LeagueResponse;
 import com.sports.server.query.dto.response.LeagueSportResponse;
 import com.sports.server.query.dto.response.LeagueTeamPlayerResponse;
-import com.sports.server.query.dto.response.LeagueTeamResponse;
 import com.sports.server.support.AcceptanceTest;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-
-import java.time.LocalDateTime;
-import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 @Sql(scripts = "/league-fixture.sql")
 public class LeagueQueryAcceptanceTest extends AcceptanceTest {
@@ -96,33 +95,6 @@ public class LeagueQueryAcceptanceTest extends AcceptanceTest {
                 () -> assertThat(actual)
                         .map(LeagueSportResponse::sportId)
                         .containsExactly(threeBuildingCup)
-        );
-    }
-
-    @Test
-    void 리그의_모든_리그팀을_조회한다() {
-        // given
-        Long threeBuildingCup = 1L;
-
-        // when
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .when()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .get("/leagues/{leagueId}/teams", threeBuildingCup)
-                .then().log().all()
-                .extract();
-
-        // then
-        List<LeagueTeamResponse> actual = toResponses(response, LeagueTeamResponse.class);
-        assertAll(
-                () -> assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value()),
-                () -> assertThat(actual)
-                        .map(LeagueTeamResponse::teamName)
-                        .containsExactly("경영 야생마", "미컴 축구생각", "서어 뻬데뻬"),
-                () -> assertThat(actual)
-                        .map(LeagueTeamResponse::leagueTeamId)
-                        .containsExactly(1L, 3L, 2L)
-
         );
     }
